xkb: new xkb implementation
Clear out the bad formatting, random greek, and other weird shit! This
is a complete rewrite of the XKB symbols file, with greek in locations
that makes sense for 3l (not Neo), and new extended unicode symbol
layers.

Everything that was in the original 3l spec should remain unchanged.

Also, previous mods moved to xkb variants (e.g., for chromebook).

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>